Report.java
Above is the code to generate custom testNG report, you can find may custom report online but this is just simple version. This custom report will include the logs logged in Reporter class, which can be used to explain and monitory the flow of each test. Instead of explaining a test is passed or failed, these logs will explain how this test was marked as passed based on the logs which is printed in the report after each test case.

You can review this code and comments to improve the code appreciated.